refactor: integerをbigintに変更

+++ db/migrate/20260316111956_add_group_id_to_posts.rb
@@ -1,6 +1,6 @@
 class AddGroupIdToPosts < ActiveRecord::Migration[6.1]
   def change
-    add_column :posts, :group_id, :bigint
+    add_column :posts, :group_id, :integer
 
     add_foreign_key :posts, :groups
